module2/requirements.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
